Everything pretty much
Added docblocks to files, generated the first rough and incomplete version of the docs at /doc Added a few of the trouble methods including:

hmyv2_getDelegationsByDelegatorByBlockNumber
hmyv2_call.
hmyv2_getStorageAt
hmyv2_estimateGas

Spent time debugging hmyv2_call. Parameters in Node API docs were way off. All numbers for this call are in hex. Found information for _call and _getStorageAt and _estimateGas and enabled them. We are at 60 of 63 API Calls now.

There is still a TON of work to do