Update to 0.9.0.2-alpha
Added ways to block by ip address as well as allow requests only specific ip addresses in case someone wanted to not allow public access to PHPH1.

Lots of documentation work

cleaned up debugging code a bit which cleaned up the methods code and also enhanced the debugging data that is presented.

created a setting to allow big data returns. It is used to turn off the ability to send page size requests with a -1 option which could potentially return huge amounts of data.

finished up the PHPH1 class docblocks for all the newly added code.