Now supports POST requests and produces simple sample javascript for post request
Lots of work, rearranged file structure a bit. Added the ability to use POST for requests and also set the explorer up to dynamically build sample javascript code for post requests. Created a post request sample in jstest.html as well. Some work on documentation but more needs to be done. Also added code highlighting to the explorer output page.