A few fixes and documentation work.
Fixed an issue with boolean detection in the explorer, Created the minimal PHPH1 class and front page documentation with code examples, just a little bit to go. Worked on the minimal package with a couple simple examples. Still have to run PHPDoc on the minimal package for its own documentation. Also validated with W3 validator and fixed any issues.